Confirmar eliminado de ficha.

+++ src/app/features/admin/views/valoracionoficial/valoracionoficial.component.ts
@@ -72,6 +72,7 @@ export class ValoracionoficialComponent implements OnInit {
 
   eliminarFicha(ficha)
   {
+    if(!confirm('¿Confirma borrado de ficha?')) return;
     this.valoracionesService.delete(ficha._id).then(response=>
     {
       this.cargarValoraciones();
